src/utils.py
def load_csv(file_path):
    import pandas as pd
    return pd.read_csv(file_path)

def save_csv(dataframe, file_path):
    dataframe.to_csv(file_path, index=False)

def preprocess_data(dataframe):
    # Implement any necessary preprocessing steps here
    return dataframe

def log(message):
    import logging
    logging.basicConfig(level=logging.INFO)
    logging.info(message)

def configure_environment(env_file):
    from dotenv import load_dotenv
    import os
    load_dotenv(env_file)
    return os.environ