requirements.txt
fastapi
pandas
numpy
scikit-learn
matplotlib
seaborn
pytest
jupyter
uvicorn
python-dotenv
scipy